refactor: getTotalDiscountForCombo 함수를 새 파일로 분리
